tests: refactor m2m rescursive fixtures
