+++ tests/assets/m2m_rescursive/_tests.py
@@ -1,23 +1,5 @@
 import pytest
 from models import Node
-from settings import TORTOISE_ORM
-from tortoise import Tortoise
-
-from aerich import Command
-
-
-@pytest.fixture(scope="session")
-def anyio_backend() -> str:
-    return "asyncio"
-
-
-@pytest.fixture(autouse=True)
-async def init_connections():
-    await Tortoise.init(TORTOISE_ORM)
-    try:
-        yield
-    finally:
-        await Command.aclose()
 
 
 @pytest.mark.anyio
