+++ tests/assets/m2m_rescursive/conftest_.py
@@ -1,0 +1,1 @@
+../conftest_.py
